+++ Modules/Combat/Events/Player/Warped/ValorWarpedEvent.cs
@@ -57,8 +57,8 @@ internal sealed class ValorWarpedEvent : WarpedEvent
         }
 
         e.Player.Write(Virtue.Valor.Name, int.MaxValue.ToString());
-        CombatModule.State.HeroQuest?.UpdateTrialProgress(Virtue.Valor);
         Game1.chatBox.addMessage(I18n.Virtues_Recognize_Yoba(), Color.Green);
+        CombatModule.State.HeroQuest?.UpdateTrialProgress(Virtue.Valor);
         this.Disable();
     }
 }