+++ Modules/Tweex/Patchers/Integration/NewForgeMenuCraftItemPatcher.cs
@@ -54,6 +54,8 @@ internal sealed class NewForgeMenuCraftItemPatcher : HarmonyPatcher
                 right_item!.ParentSheetIndex == ObjectIds.GlowRing:
                 __result = new Ring(ObjectIds.GlowstoneRing);
                 break;
+            default:
+                return true; // run original logic
         }
 
         return false; // run original logic